+++ lib/runners/config.rb
@@ -1,16 +1,26 @@
 module Runners
   class Config
     class Error < UserError
+      attr_reader :path_name
       attr_reader :raw_content
 
-      def initialize(message, raw_content)
+      def initialize(message, path_name:, raw_content:)
         super(message)
+        @path_name = path_name
         @raw_content = raw_content
       end
     end
 
     class BrokenYAML < Error; end
-    class InvalidConfiguration < Error; end
+
+    class InvalidConfiguration < Error
+      attr_reader :attribute
+
+      def initialize(message, path_name:, raw_content:, attribute:)
+        super(message, path_name: path_name, raw_content: raw_content)
+        @attribute = attribute
+      end
+    end
 
     FILE_NAME = "sider.yml".freeze
     FILE_NAME_OLD = "sideci.yml".freeze
@@ -109,8 +119,10 @@ module Runners
       begin
         YAML.safe_load(yaml, symbolize_names: true, filename: path_name)
       rescue Psych::SyntaxError => exn
-        message = "`#{exn.file}` is broken at line #{exn.line} and column #{exn.column}"
-        raise BrokenYAML.new(message, yaml)
+        raise BrokenYAML.new(
+          "`#{exn.file}` is broken at line #{exn.line} and column #{exn.column}",
+          path_name: path_name, raw_content: yaml,
+        )
       end
     end
 
@@ -119,13 +131,17 @@ module Runners
     def check_schema(object)
       object ? Schema::Config.payload.coerce(object) : {}
     rescue StrongJSON::Type::UnexpectedAttributeError => exn
-      attr = "#{exn.path}.#{exn.attribute}".delete_prefix("$.")
-      message = "The attribute `#{attr}` in your `#{path_name}` is unsupported. Please fix and retry."
-      raise InvalidConfiguration.new(message, raw_content!)
+      attr = [exn.path, exn.attribute].join(".")
+      raise InvalidConfiguration.new(
+        "`#{attr.delete_prefix('$.')}` in `#{path_name}` is unsupported",
+        path_name: path_name, raw_content: raw_content || "", attribute: attr,
+      )
     rescue StrongJSON::Type::TypeError => exn
-      attr = exn.path.to_s.delete_prefix("$.")
-      message = "The value of the attribute `#{attr}` in your `#{path_name}` is invalid. Please fix and retry."
-      raise InvalidConfiguration.new(message, raw_content!)
+      attr = exn.path.to_s
+      raise InvalidConfiguration.new(
+        "`#{attr.delete_prefix('$.')}` value in `#{path_name}` is invalid",
+        path_name: path_name, raw_content: raw_content || "", attribute: attr,
+      )
     end
   end
 end