+++ test/smokes/checkstyle/expectations.rb
@@ -138,8 +138,7 @@ s.add_test(
 s.add_test(
   "broken_sideci_yml",
   type: "failure",
-  message:
-    "The value of the attribute `linter.checkstyle.exclude` in your `sideci.yml` is invalid. Please fix and retry.",
+  message: "`linter.checkstyle.exclude` value in `sideci.yml` is invalid",
   analyzer: :_
 )