+++ test/smokes/eslint/expectations.rb
@@ -162,8 +162,7 @@ s.add_test(
 s.add_test(
   "broken_sideci_yml",
   type: "failure",
-  message:
-    "The value of the attribute `linter.eslint.npm_install` in your `sideci.yml` is invalid. Please fix and retry.",
+  message: "`linter.eslint.npm_install` value in `sideci.yml` is invalid",
   analyzer: :_
 )