+++ test/smokes/phpmd/expectations.rb
@@ -125,7 +125,7 @@ s.add_test(
   "broken_sideci_yml",
   type: "failure",
   message:
-    "The value of the attribute `linter.phpmd.minimumpriority` in your `sideci.yml` is invalid. Please fix and retry.",
+    "`linter.phpmd.minimumpriority` value in `sideci.yml` is invalid",
   analyzer: :_
 )