+++ test/smokes/rubocop/expectations.rb
@@ -218,8 +218,7 @@ s.add_test(
   "broken_sideci_yml",
   type: "failure",
   analyzer: :_,
-  message:
-    "The value of the attribute `linter.rubocop.gems[0]` in your `sideci.yml` is invalid. Please fix and retry."
+  message: "`linter.rubocop.gems[0]` value in `sideci.yml` is invalid"
 )
 
 s.add_test(