+++ test/smokes/scss_lint/expectations.rb
@@ -140,7 +140,6 @@ s.add_test(
 s.add_test(
   "broken_sideci_yml",
   type: "failure",
-  message:
-    "The value of the attribute `linter.scss_lint.config` in your `sideci.yml` is invalid. Please fix and retry.",
+  message: "`linter.scss_lint.config` value in `sideci.yml` is invalid",
   analyzer: :_
 )