+++ test/smokes/swiftlint/expectations.rb
@@ -176,8 +176,7 @@ s.add_test(
 s.add_test(
   "broken_sideci_yml",
   type: "failure",
-  message:
-    "The value of the attribute `linter.swiftlint.lenient` in your `sideci.yml` is invalid. Please fix and retry.",
+  message: "`linter.swiftlint.lenient` value in `sideci.yml` is invalid",
   analyzer: :_
 )